requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
bcrypt==3.2.0
dash-html-components==2.0.0
dash-renderer==1.9.1
Werkzeug==2.1.1
argparse